+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 exceptional situation.
 - [Cache] Added various cache mechanisms inside `cache` module: `EnfoldCache`. `AllowanceCache`.
 - [Policy] Added `PolicyAllow` and `PolicyDeny` for more convenient Policy effects declaration.
+- [Storage] Added `storage.observable.ObservableMutationStorage` as a Storage whose modify interface is observable.
 
 ### Changed
 - [Inquiry] Inquiry objects equality is now based on their contents equality. Same for its hash value.